SVG diagram improvements
#1039
