+++ waltz-ng/client/svg-diagram/component/svg-diagrams.html
@@ -1,0 +1,18 @@
+<div>
+    <uib-tabset>
+        <uib-tab heading='{{ diagram.name }}'
+                 ng-repeat="diagram in $ctrl.diagrams | orderBy:'priority'"
+                 deselect="$ctrl.hide(diagram)"
+                 select="$ctrl.show(diagram)">
+            <div ng-if="diagram.visible">
+                <waltz-svg-diagram diagram="diagram"
+                                   block-processor="$ctrl.blockProcessor">
+                </waltz-svg-diagram>
+            </div>
+            <hr class="waltz-hr-dotted">
+            <div class="text-muted"
+                 ng-bind="diagram.description">
+            </div>
+        </uib-tab>
+    </uib-tabset>
+</div>
