+++ waltz-ng/client/svg-diagram/index.js
@@ -2,8 +2,8 @@
 export default (module) => {
     module
         .service('SvgDiagramStore', require('./services/svg-diagram-store'))
-        .directive('waltzSvgDiagram', require('./directives/svg-diagram'))
-        .directive('waltzSvgDiagrams', require('./directives/svg-diagrams'))
+        .component('waltzSvgDiagram', require('./component/svg-diagram'))
+        .component('waltzSvgDiagrams', require('./component/svg-diagrams'))
         .component('waltzCommonSvgDefs', require('./component/common-svg-defs'));
 
 };
