+++ waltz-ng/index.template.html
@@ -30,13 +30,62 @@
             <!-- Content -->
             <div ui-view="content"
                  class="waltz-content col-xs-12">
-                <img style="
+
+                <div style="
                         position: fixed;
                         left: 50%;
                         top: 50%;
                         margin-top: -122px;
-                        margin-left: -150px;"
-                     src="images/branding/waltz_icon+text_positive_bw_300px.png">
+                        margin-left: -150px;">
+                    <?xml version="1.0" encoding="utf-8"?>
+                    <svg version="1.1"
+                     id="Layer_1"
+                     xmlns="http://www.w3.org/2000/svg"
+                     xmlns:xlink="http://www.w3.org/1999/xlink"
+                     x="0px"
+                     y="0px"
+                     width="300px"
+                     height="245px"
+                     viewBox="0 0 600 490"
+                     xml:space="preserve">
+                    <g>
+                        <g>
+                            <path d="M35.7,461.7L65.8,387L96,461.7l24.5-69.8h11.2l-35.7,98l-30.1-75l-30.1,75L0,391.9h11.2L35.7,461.7z"/>
+                            <path d="M193.6,458.6L182.3,485H171l43-97.6l41.9,97.6h-11.5l-11-26.4H193.6z M213.7,411.6l-15.9,37.3h31.5L213.7,411.6z"/>
+                            <path d="M325.7,475.4H372v9.6h-56.7v-93.1h10.4L325.7,475.4L325.7,475.4z"/>
+                            <path d="M451.7,485h-10.4v-83.5h-27.6v-9.6h65.9v9.6h-27.9V485z"/>
+                            <path d="M598.6,475.4v9.6h-66.7l50.8-83.5h-44.3v-9.6H600l-50.9,83.5H598.6z"/>
+                        </g>
+                        <g>
+                            <path d="M147,270.8L62.9,115.4c-2.2,1.8-4.7,3.1-7.4,4l84.1,155.4C141.8,273,144.3,271.7,147,270.8z"/>
+                            <path d="M322.2,148.8L167.9,274.9c2.2,1.8,4,4,5.3,6.5l154.3-126.1C325.4,153.5,323.6,151.3,322.2,148.8z"/>
+                            <path d="M409.2,231.2L357.4,154c-2,2-4.4,3.6-7,4.7l51.9,77.2C404.3,233.9,406.6,232.3,409.2,231.2z"/>
+                            <path d="M542.9,42.8L425.8,230.9c2.7,1,5.1,2.5,7.1,4.5L550,47.2C547.4,46.2,545,44.7,542.9,42.8z"/>
+                            <g>
+                                <path d="M48.8,80.3c9.9,0,18,8.1,18,18s-8.1,18-18,18s-18-8.1-18-18S38.9,80.3,48.8,80.3 M48.8,71.9c-14.6,0-26.4,11.8-26.4,26.4
+                                    s11.8,26.4,26.4,26.4s26.4-11.8,26.4-26.4S63.4,71.9,48.8,71.9L48.8,71.9z"/>
+                            </g>
+                            <g>
+                                <path d="M153.7,274c9.9,0,18,8.1,18,18s-8.1,18-18,18s-18-8.1-18-18S143.8,274,153.7,274 M153.7,265.6
+                                    c-14.6,0-26.4,11.8-26.4,26.4s11.8,26.4,26.4,26.4s26.4-11.8,26.4-26.4C180.1,277.4,168.3,265.6,153.7,265.6L153.7,265.6z"/>
+                            </g>
+                            <g>
+                                <path d="M341.7,120.3c9.9,0,18,8.1,18,18s-8.1,18-18,18s-18-8.1-18-18S331.8,120.3,341.7,120.3 M341.7,111.9
+                                    c-14.6,0-26.4,11.8-26.4,26.4c0,14.6,11.8,26.4,26.4,26.4s26.4-11.8,26.4-26.4S356.3,111.9,341.7,111.9L341.7,111.9z"/>
+                            </g>
+                            <g>
+                                <path d="M417.9,233.6c9.9,0,18,8.1,18,18c0,9.9-8.1,18-18,18s-18-8.1-18-18C399.9,241.7,408,233.6,417.9,233.6 M417.9,225.2
+                                    c-14.6,0-26.4,11.8-26.4,26.4c0,14.6,11.8,26.4,26.4,26.4s26.4-11.8,26.4-26.4C444.3,237,432.5,225.2,417.9,225.2L417.9,225.2z"
+                                    />
+                            </g>
+                            <g>
+                                <path d="M558,8.6c9.9,0,18,8.1,18,18s-8.1,18-18,18s-18-8.1-18-18S548.1,8.6,558,8.6 M558,0.2c-14.6,0-26.4,11.8-26.4,26.4
+                                    S543.4,53,558,53c14.6,0,26.4-11.8,26.4-26.4S572.6,0.2,558,0.2L558,0.2z"/>
+                            </g>
+                        </g>
+                    </g>
+                    </svg>
+                </div>
             </div>
         </div>
     </div>
