+++ waltz-ng/webpack.config.js
@@ -37,7 +37,7 @@ module.exports = {
             { test: /\.scss$/, loader: 'style!css!sass' },
             { test: /\.css$/, loader: 'style!css' },
             { test: /\.(ttf|eot|svg)(\?v=[0-9]\.[0-9]\.[0-9])?$/, loader: 'url-loader?limit=8192' },
-            { test: /\.png$/, loader: 'url-loader?mimetype=image/png&limit=8192' },
+            { test: /\.png$/, loader: 'url-loader?mimetype=image/png&limit=16384' },
             { test: /\.html?$/, loader: 'html-loader' },
             { test: /\.woff(\?v=[0-9]\.[0-9]\.[0-9])?$/, loader: 'url-loader?limit=8192&minetype=application/font-woff' },
             { test: /\.woff2(\?v=[0-9]\.[0-9]\.[0-9])?$/, loader: 'url-loader?limit=8192&minetype=application/font-woff2' }
